Add ability to fingerprint patient golden ID and reprocess on golden ID changes
This change add a reprocess mediator that accepts request to reprocess events associated with a particular golden ID. Logstash has been updated to enrich certain events with a golden ID field as well as to detect updates to Golden Patient Records. When an update is detected a request is send to the reprocess mediator to reprocess events associated with that golden ID. The reprocess mediator queries the raw event from elastic search and sendsthem to the logstash enrich pipeline for reprocessing.